# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MCP Host Admin API (Stub)
  version: 0.1.5
servers:
  - url: http://localhost:8080
paths:
...
  /prompt-templates:
    get:
      summary: Listet Prompt-Vorlagen des aktuellen Nutzers auf
      parameters:
        - in: query
          name: scope
          schema:
            type: string
            enum: [task, agent, global]
        - in: query
          name: target_id
          schema:
            type: string
        - in: query
          name: includeGlobal
          schema:
            type: boolean
      responses:
        '200':
          description: Prompt-Vorlagen
    post:
      summary: Legt eine neue Prompt-Vorlage an
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scope:
                  type: string
                  enum: [task, agent, global]
                target_id:
                  type: string
                title:
                  type: string
                content:
                  type: string
              required: [scope, content]
      responses:
        '201':
          description: Vorlage erstellt
  /prompt-templates/{id}:
    put:
      summary: Aktualisiert eine bestehende Prompt-Vorlage
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                title:
                  type: string
              required: [content]
      responses:
        '200':
          description: Vorlage aktualisiert
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  scope:
                    type: string
                  target_id:
                    type: string
                    nullable: true
                  title:
                    type: string
                  content:
                    type: string
                  created_at:
                    type: string
                    format: date-time
                  updated_at:
                    type: string
                    format: date-time
        '400':
          description: Ungültige Eingabe (invalid_argument)
        '404':
          description: Vorlage nicht gefunden
    delete:
      summary: Löscht eine Prompt-Vorlage
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Vorlage gelöscht
        '404':
          description: Vorlage nicht gefunden
...